implementation of word

+++ mpp-cpu/src/lib/electronic/multiplexer.h
@@ -8,7 +8,7 @@
 #ifndef multiplexer_h
 #define multiplexer_h
 
-typedef unsigned long long MXInput;
+typedef int MXInput;
 
 MXInput run_4x1_mx(unsigned int sel, MXInput i1, MXInput i2, MXInput i3, MXInput i4);
 MXInput run_2x1_mx(unsigned int sel, MXInput i1, MXInput i2);
